fix(drawer): pr问题修改
