+++ packages/zent/src/drawer/components/DrawerContent/index.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { useMemo } from 'react';
+import { useMemo, useCallback, useRef } from 'react';
 import cx from 'classnames';
 import { CSSTransition } from 'react-transition-group';
 
@@ -8,6 +8,7 @@ import { TransitionTimeOut, StyleMap } from '../../constants';
 import { renderHeader } from './Header';
 import { renderFooter } from './Footer';
 import { renderCloseBtn } from './CloseBtn';
+import { addEventListener } from '../../../utils/component/event-handler';
 
 const DrawerContent: React.FC<IDrawerContent> = ({
   mask,
@@ -18,23 +19,37 @@ const DrawerContent: React.FC<IDrawerContent> = ({
   placement,
   closeBtn,
   onClose,
-  onEntered,
   onExited,
-  onExit,
   width,
   height,
 }) => {
+  const refEventer = useRef<() => void>();
+  const drawerEl = useRef(null);
+
+  const onDrawerOpened = useCallback(() => {
+    if (!mask) {
+      refEventer.current = addEventListener(document, 'click', e => {
+        if (!drawerEl.current?.contains(e.target)) {
+          onClose();
+        }
+      });
+    }
+  }, [onClose, mask]);
+
+  const onDrawerExit = useCallback(() => {
+    if (!mask) {
+      refEventer.current?.();
+    }
+  }, [mask]);
+
   const customWH = useMemo(() => {
     const handleNumber = data =>
       typeof data === 'number' ? `${data}px` : data;
 
-    if (('left' === placement || 'right' === placement) && width) {
+    if ('left' === placement || 'right' === placement) {
       return { width: handleNumber(width) };
     }
-    if (height) {
-      return { height: handleNumber(height) };
-    }
-    return {};
+    return { height: handleNumber(height) };
   }, [placement, width, height]);
 
   return (
@@ -45,19 +60,15 @@ const DrawerContent: React.FC<IDrawerContent> = ({
       in={visible}
       timeout={TransitionTimeOut}
       classNames={`zent-drawer-transition-${placement}`}
-      onEntered={onEntered}
-      onExit={onExit}
+      onEntered={onDrawerOpened}
+      onExit={onDrawerExit}
       onExited={onExited}
     >
       <div
+        ref={drawerEl}
         className={cx('zent-drawer-content', {
           [`zent-drawer-content--transparent`]: !mask,
         })}
-        onClick={e => {
-          if (!mask) {
-            e.nativeEvent.stopImmediatePropagation();
-          }
-        }}
         style={{
           ...StyleMap[placement],
           ...customWH,
